Bug fixes related to running  replicates with spatially varying boundaries
Changelog:

- Add Init function in FEM_Common class to initialize parameters in FEM_Common
- Call above Init function within InitPhysical in solver.cpp when the mesh
  parameters for a new replicate are refreshed. This wipes the slate clean for
  a new replicate
- With the above Init function in place, the initialization of domains after
  pull-in can be removed, since the domains will get re-initialized when a new
  replicate is started
- Fix bug in stop criterion for dynamic analysis loop in psolver.py by properly
  testing whether the returned value is None. Previous condition had a bug where
  the loop stops if the return value is a float of value zero
- Fix bugs in renaming/deletion of cache/backup files in ParametricSolver

Code has been tested to run dynamic analysis on multiple replicates with
spatially varying boundaries

+++ unyquesolver/psolver.py
@@ -161,7 +161,7 @@ class ParametricSolver(object):
             # as tuples with the current time as the first item, followed by the
             # post-processed results. Continue solving until a None is returned
             # as the second item
-            while rvalue[1]:
+            while rvalue[1] is not None:
 
                 # Store previously computed solution
                 solution.append(rvalue)
@@ -171,7 +171,8 @@ class ParametricSolver(object):
                 if self.use_cache and counter % cache_point == 0:
 
                     # Backup existing cache file in case caching is interrupted
-                    os.rename(cache_file, cache_file + '.bak')
+                    if os.path.exists(cache_file):
+                        os.rename(cache_file, cache_file + '.bak')
 
                     # Write to cache file
                     with open(cache_file, 'wb') as f:
@@ -181,7 +182,10 @@ class ParametricSolver(object):
                                       self._solver.fluid_domain), f)
 
                     # Delete backup
-                    os.remove(cache_file + '.bak')
+                    try:
+                        os.remove(cache_file + '.bak')
+                    except OSError:
+                        pass
 
                 # Run solver at new time step
                 rvalue = self._solver.Solve(pset)
